Large change to main and insert
Changed the way main looked, although it should have the same kind of
functionality as it did before. I thought it would be kind of neat to
use an iterator to go through the transactions list. Used switch
statement to do the "if-else-if-else.." for each type of transaction.
Put in skeletons for the search, delete, update, and print methods and
added the shopping cart input argument for each of them. Wrote insert(),
but haven't thoroughly tested it yet.