update() method and change to printAttribute
Filled out the update() method and it seems to work on some small tests.
The wording in printAttribute() for each class was changed a bit to make
it easier to read. There's also some tester code in the main while loop
that I commented out that will print out the contents of the shopping
cart.