Error check decimal / zero fix
Added a block in the insert case in error check that checks for if there
is a decimal and if the fraction value is all zero. If that is the case
then it puts the integer part back as a String into that spot on tokens.
This means that the check for tokens[4] and tokens[5] was removed from
each grocery/clothing/electronics checks, meaning check could be flagged
as true, however there are new try/catches in this new block to prevent
an invalid input and reflag it as false if it fails.